Split tests in more files, add wmd test suite and increased test coverage
